Move friend request btn to organisms
